Refactored elements of solution
