+++ src/functions/pollJournals/framework/databases/mysql/__mocks__/users.mock.ts
@@ -1,0 +1,8 @@
+import {ExaminerRecord} from '../../../../domain/examiner-record';
+
+export const TARSUserMock: ExaminerRecord[] = [
+  {individual_id: 1, staff_number: '2'},
+  {individual_id: 3, staff_number: '4'},
+  {individual_id: 5, staff_number: '6'},
+  {individual_id: 7, staff_number: '8'},
+];
