+++ src/functions/pollJournals/framework/databases/mysql/advance-test-slots-repository.ts
@@ -1,10 +1,10 @@
 import * as mysql from 'mysql2';
 import * as moment from 'moment';
 import {AdvanceTestSlotRow, mapRow} from './row-mappers/advance-test-slot-row-mapper';
-import {poolQuery, query} from '../../../../../common/framework/mysql/database';
+import {poolQuery} from '../../../../../common/framework/mysql/database';
 import {info, customDurationMetric} from '@dvsa/mes-microservice-common/application/utils/logger';
 import {ExaminerAdvanceTestSlot} from '../../../domain/examiner-advance-test-slot';
-import {formatDateToStartTime} from '../../../application/formatters/date-formatter';
+import {TARSAdvanceTestSlotMock} from './__mocks__/advance-test-slot.mock';
 
 /**
  * Get all test slots in the advanced time window.
@@ -18,72 +18,7 @@ export const getAdvanceTestSlots = async (
   connectionPool: mysql.Pool, startDate: Date, nextWorkingDay: Date,
   daysRange: number): Promise<ExaminerAdvanceTestSlot[]> => {
   if (process.env.IS_DEV) {
-    return [
-      {
-        examinerId: 1,
-        advanceTestSlot: {
-          slotDetail: {
-            slotId: 200,
-            start: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-            duration: 57,
-          },
-          testCentre: {
-            centreId: 1,
-            centreName: 'Test Centre 1',
-            costCode: 'TC1',
-          },
-          vehicleTypeCode: 'VT1',
-        },
-      },
-      {
-        examinerId: 3,
-        advanceTestSlot: {
-          slotDetail: {
-            slotId: 201,
-            start: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-            duration: 57,
-          },
-          testCentre: {
-            centreId: 2,
-            centreName: 'Test Centre 2',
-            costCode: 'TC2',
-          },
-          vehicleTypeCode: 'VT2',
-        },
-      },
-      {
-        examinerId: 5,
-        advanceTestSlot: {
-          slotDetail: {
-            slotId: 202,
-            start: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-            duration: 57,
-          },
-          testCentre: {
-            centreId: 3,
-            centreName: 'Test Centre 3',
-            costCode: 'TC3',
-          },
-          vehicleTypeCode: 'VT3',
-        },
-      },
-      {
-        examinerId: 7,
-        advanceTestSlot: {
-          slotDetail: {
-            slotId: 203,
-            start: moment(Date.now()).format('YYYY-MM-DD HH:mm:ss'),
-            duration: 57,
-          },
-          testCentre: {
-            centreId: 4,
-            centreName: 'Test Centre 4',
-            costCode: 'TC4',
-          },
-          vehicleTypeCode: 'VT4',
-        },
-      },
-    ];
+    return TARSAdvanceTestSlotMock;
   }
 
   const sqlYearFormat = 'YYYY-MM-DD';
