+++ src/functions/pollJournals/framework/databases/mysql/personal-commitment-repository.ts
@@ -4,6 +4,7 @@ import { mapRow, PersonalCommitmentRow } from './row-mappers/personal-commitment
 import { poolQuery } from '../../../../../common/framework/mysql/database';
 import { info, customDurationMetric } from '@dvsa/mes-microservice-common/application/utils/logger';
 import { ExaminerPersonalCommitment } from '../../../domain/examiner-personal-commitment';
+import {TARSPersonalCommitmentMock} from './__mocks__/personal-commitment.mock';
 
 /**
  * Get all personal commitments, for the specified time window.
@@ -15,44 +16,7 @@ import { ExaminerPersonalCommitment } from '../../../domain/examiner-personal-co
 export const getPersonalCommitments = async (connectionPool: mysql.Pool, startDate: Date, durationDays: number):
 Promise<ExaminerPersonalCommitment[]> => {
   if (process.env.IS_DEV) {
-    return [
-      {
-        examinerId: 1,
-        personalCommitment: {
-          commitmentId: 1,
-          slotId: 1,
-          activityCode: '1',
-          activityDescription: 'Personal Commitment 1',
-        },
-      },
-      {
-        examinerId: 3,
-        personalCommitment: {
-          commitmentId: 2,
-          slotId: 2,
-          activityCode: '2',
-          activityDescription: 'Personal Commitment 2',
-        },
-      },
-      {
-        examinerId: 5,
-        personalCommitment: {
-          commitmentId: 3,
-          slotId: 3,
-          activityCode: '3',
-          activityDescription: 'Personal Commitment 3',
-        },
-      },
-      {
-        examinerId: 7,
-        personalCommitment: {
-          commitmentId: 4,
-          slotId: 4,
-          activityCode: '4',
-          activityDescription: 'Personal Commitment 4',
-        },
-      },
-    ];
+    return TARSPersonalCommitmentMock;
   }
   const windowStart = moment(startDate);
   const windowEnd = windowStart.clone().add({days: durationDays}).subtract({seconds: 1});
